Store non-fs items data into dav_properties table
