+++ apps/dav/appinfo/info.xml
@@ -5,7 +5,7 @@
 	<description>ownCloud WebDAV endpoint</description>
 	<licence>AGPL</licence>
 	<author>owncloud.org</author>
-	<version>0.3.1</version>
+	<version>0.3.2</version>
 	<default_enable/>
 	<use-migrations>true</use-migrations>
 	<types>
